requirements.txt
requests>=2.28.1
selenium>=4.7.2
python-dateutil>=2.8.2
flask>=2.2.0